expense_tracker_backend/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
drf-yasg==1.21.10
flake8==7.2.0
inflection==0.5.1
iniconfig==2.1.0
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
pycodestyle==2.13.0
pyflakes==3.3.2
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.9
# Optional helper if present; not required. If installed, DATABASE_URL parsing will be automatic.
dj-database-url==2.3.0
djangorestframework-simplejwt==5.3.1
